app/requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.2.0
astrapy==0.6.2
attrs==23.2.0
backoff==2.2.1
beautifulsoup4==4.12.2
bs4==0.0.1
cassandra-driver==3.29.0
cassio==0.1.4
certifi==2023.11.17
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
dataclasses-json==0.6.3
Deprecated==1.2.14
distro==1.9.0
emoji==2.9.0
filetype==1.2.0
frozenlist==1.4.1
fsspec==2023.12.2
geomet==0.2.1.post1
greenlet==3.0.3
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.2
httpx==0.25.2
hyperframe==6.0.1
idna==3.6
joblib==1.3.2
jsonpatch==1.33
jsonpointer==2.4
langchain==0.0.354
langchain-community==0.0.11
langchain-core==0.1.9
langdetect==1.0.9
langsmith==0.0.81
llama-index==0.9.29
lxml==5.1.0
marshmallow==3.20.2
multidict==6.0.4
mypy-extensions==1.0.0
nest-asyncio==1.5.9
networkx==3.2.1
nltk==3.8.1
numpy==1.26.3
openai==1.8.0
packaging==23.2
pandas==2.1.4
pydantic==2.5.3
pydantic_core==2.14.6
pypdf==3.17.4
python-dateutil==2.8.2
python-dotenv==1.0.0
python-iso639==2024.1.2
python-magic==0.4.27
pytz==2023.3.post1
PyYAML==6.0.1
ragstack-ai==0.5.0
rapidfuzz==3.6.1
regex==2023.12.25
requests==2.31.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==2.0.25
tabulate==0.9.0
tenacity==8.2.3
tiktoken==0.5.2
tqdm==4.66.1
typing-inspect==0.9.0
typing_extensions==4.9.0
tzdata==2023.4
unstructured==0.10.30
urllib3==2.1.0
wrapt==1.16.0
yarl==1.9.4
